Cambios definitivos en back
Arreglo de main , limpieza de funciones...
