+++ datos.txt
@@ -1,6 +1,6 @@
-20
-0
+30
 1
+0
 1
 0
 1
