+++ empleados_mas_incumplidores.csv
@@ -1,21 +1,31 @@
 empleado_id,faltas_acumuladas,faltas_seguidas,falta_lunes_viernes,llegada_tarde,retiro_temprano,anomaly_score
-68,31,5,19,1,1,-0.19082371428565748
-70,18,6,8,3,4,-0.17585314380905448
-32,25,4,20,5,2,-0.1633103509841335
-16,11,1,4,3,10,-0.13386475188097513
-18,27,3,19,5,1,-0.12490193056418852
-87,12,3,5,3,9,-0.07634750066197415
-86,4,1,0,4,6,-0.07310123351717479
-26,16,4,8,5,1,-0.04947028391266095
-4,11,1,6,6,0,-0.04454285014593673
-76,18,3,5,2,8,-0.04260606019541868
-52,21,3,14,2,3,-0.030364533667462612
-81,8,1,3,2,1,-0.02478466358528586
-62,11,1,0,3,3,-0.02319580302168678
-79,22,4,10,3,4,-0.021318877629040167
-98,9,2,4,3,8,-0.018201540553220097
-14,10,1,2,1,6,-0.01234753071717376
-92,17,4,10,0,3,-0.007116028019482634
-83,9,1,1,5,4,-0.006245004347453897
-35,14,1,9,3,3,-0.00010697192743358563
-57,13,4,4,4,4,-1.7009503322951147e-05
+68,31,5,19,1,1,-0.21347828723360135
+18,27,3,19,5,1,-0.1757141987270498
+32,25,4,20,5,2,-0.16348403377539472
+86,4,1,0,4,6,-0.1409233817485997
+30,31,3,10,0,2,-0.13581878574693412
+16,11,1,4,3,10,-0.09423827853245553
+52,21,3,14,2,3,-0.08819240320268046
+84,26,3,6,3,6,-0.08074377630480006
+9,17,2,4,4,0,-0.062367834786679455
+87,12,3,5,3,9,-0.058187258176812306
+79,22,4,10,3,4,-0.05590841225034365
+4,11,1,6,6,0,-0.0549309133493458
+23,8,2,1,2,4,-0.052104876824702606
+96,17,2,6,1,0,-0.047173013551181364
+98,9,2,4,3,8,-0.04182775476189976
+76,18,3,5,2,8,-0.03283920800536716
+5,7,1,3,5,5,-0.027299780855771394
+97,18,2,7,5,6,-0.022868834793736192
+14,10,1,2,1,6,-0.01857268341916435
+3,9,2,2,6,3,-0.017959099754700614
+81,8,1,3,2,1,-0.013522237199205533
+62,11,1,0,3,3,-0.012126338912867074
+88,18,2,10,1,4,-0.009981033489886604
+29,9,3,3,2,1,-0.008494370968124498
+92,17,4,10,0,3,-0.00598287948872106
+95,10,2,2,3,4,-0.005619984199145223
+85,6,1,4,1,3,-0.003997645616704182
+83,9,1,1,5,4,-0.0025573671549455335
+49,19,3,8,5,5,-0.002224978980487846
+6,15,3,7,1,5,-0.00024067154696277848
