+++ generador de dataset/agrega_anomalias.py
